+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/exporter.py
@@ -28,7 +28,7 @@ class BaseExporter(object):
         self.max_batch_size = options.max_batch_size
         # TODO: queue should be moved to tracer
         # too much refactor work, leave to the next PR
-        self._queue = Queue(capacity=8192)  # TODO: make this configurable
+        self._queue = Queue(capacity=options.queue_capacity)
         # TODO: worker should not be created in the base exporter
         self._worker = Worker(self._queue, self)
         self._worker.start()
@@ -61,7 +61,9 @@ class Worker(threading.Thread):
         self.src = src
         self.dst = dst
         self._stopping = False
-        super(Worker, self).__init__()
+        super(Worker, self).__init__(
+            name="AzureExporter Worker"
+        )
 
     def run(self):  # pragma: NO COVER
         # Indicate that this thread is an exporter thread.