+++ contrib/opencensus-ext-httplib/CHANGELOG.md
@@ -8,7 +8,7 @@ Released 2020-02-03
 - Added `component` span attribute
 
 ## 0.7.2
-Released 2019-08-06
+Released 2019-08-26
 
 - Updated `http.status_code` attribute to be an int.
   ([#755](https://github.com/census-instrumentation/opencensus-python/pull/755))